refactor: styled.js 파일 네이밍 변경
